requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
backcall=0.2.0=pypi_0
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py36h9f0ad1d_0
cffi=1.14.0=py36hd463f26_0
cudatoolkit=10.2.89=hfd86e86_1
decorator=4.4.2=pypi_0
freetype=2.10.2=he06d7ca_0
h5py=2.10.0=pypi_0
ipdb=0.13.3=pypi_0
ipython=7.15.0=pypi_0
ipython-genutils=0.2.0=pypi_0
jedi=0.17.1=pypi_0
jpeg=9d=h516909a_0
ld_impl_linux-64=2.34=h53a641e_5
libblas=3.8.0=16_openblas
libcblas=3.8.0=16_openblas
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.5.0=hdf63c60_6
liblapack=3.8.0=16_openblas
libopenblas=0.3.9=h5ec1e0e_0
libpng=1.6.37=hed695b0_1
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc7e4089_6
libwebp-base=1.1.0=h516909a_3
llvm-openmp=10.0.0=hc9558a2_0
lz4-c=1.9.2=he1b5a44_1
medpy=0.4.0=pypi_0
mkl=2020.1=219
ncurses=6.1=hf484d3e_1002
nibabel=3.1.0=pypi_0
ninja=1.10.0=hc9558a2_0
numpy=1.18.5=py36h7314795_0
olefile=0.46=py_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=h516909a_0
packaging=20.4=pypi_0
parso=0.7.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.1.2=py36h8328e55_0
pip=20.1.1=py_1
prompt-toolkit=3.0.5=pypi_0
ptyprocess=0.6.0=pypi_0
pycparser=2.20=py_0
pygments=2.6.1=pypi_0
pyparsing=2.4.7=pypi_0
python=3.6.10=h8356626_1011_cpython
python_abi=3.6=1_cp36m
pytorch=1.0.0=py3.6_cuda9.0.176_cudnn7.4.1_1
readline=8.0=hf8c457e_0
scipy=1.5.0=pypi_0
setuptools=47.3.1=py36h9f0ad1d_0
simpleitk=1.2.4=pypi_0
six=1.15.0=pyh9f0ad1d_0
sqlite=3.30.1=hcee41ef_0
termcolor=1.1.0=pypi_0
tk=8.6.10=hed695b0_0
torchsummary=1.5.1=pypi_0
torchvision=0.2.2=py_3
tqdm=4.46.1=pypi_0
traitlets=4.3.3=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.34.2=py_1
xz=5.2.5=h516909a_0
zlib=1.2.11=h516909a_1006
zstd=1.4.4=h6597ccf_3
